+++ tests/mocks/SharedCache.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Apricity\tests\mocks;
+
+use Apricity\MicroCache;
+
+class SharedCache extends MicroCache
+{
+}