+++ tests/mocks/UniqueCache.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Apricity\tests\mocks;
+
+use Apricity\MicroCache;
+
+class UniqueCache extends MicroCache
+{
+    protected static array $microCache = [];
+}